requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Naked==0.1.31
pycryptodomex==3.10.1
pypiwin32==223
pywin32==300
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
urllib3==1.26.4
